iwg20m: add support for this board
This patch adds support for the iwg20m board to isar-cip-core.

- Kernel version notes

This board is only supported by the 4.4 kernel. For that
reason, the preferred version is set to 4.4 and the user
doesn't need to specify :opt-4.4.yml. To be precise, the
kernel 4.19 also works but Renesas is only committed to
support 4.4.

- U-boot notes

The u-boot version shipped with the board is rather
old (2013.01) and comes without distro boot support.
We could try building it with CONFIG_CMD_SOURCE to use
a boot.scr script. However, the build gave several
errors probably caused by incompatibilities with
modern compilers. Additionally, u-boot resides
on an SPI device and any failures while updating
it will brick the board. For all these reasons,
this patch relies on the u-boot shipped with the
board completely. The user needs to setup some
additional u-boot environment variables as described
in doc/README_iwg20m.md

- Kernel configuration notes

The configuration is based on the kernel's
shmobile_defconfig. Then it adds some
configuration bits for booting on an SDCard with
EXT4, booting from a ramdisk, and running Systemd.

- Bmap notes

The document doc/README_iwg20m.md uses bmaptool to
flash the image onto an SDCard.

Signed-off-by: Daniel Sangorrin <daniel.sangorrin@toshiba.co.jp>