python3-shtab: add export to the DEB_BUILD_PROFILES
When we build this recipes in downstream layer the DEB_BUILD_PROFILES not
able to find by rule file hence it is not set the required build
profiles.

Here we disable the tests `PYBUILD_DISABLE=test`, but due to this
unavailability of DEB_BUILD_PROFILES in rule file it is building along
with the test hence it's getting failed like "Not able to import module
pytest"

Exported this one to access from rule file.

Signed-off-by: Srinuvasan A <srinuvasan_a@mentor.com>