customizations: comment why we can not actually change home easily
usermod --move-home seems like the cleanest solution, or even putting
that into ksd and using USER_root[home] = "/home/root"
But that unfortunately does not work. Some sed on /etc/passwd is likely
a bad idea so leave it a symlink but add some more info as commit
message and comment.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>