only copy "A" version of rootfs part for ab-rootfs
For swupdate images, we currently copy the rootfs to both the A and the
B partition. However, the content of the B rootfs will never be used, as
the first boot always chooses the A partition. And after an swupdate
image update is performed, the content of the B version will be
overwritten.

By not copying the squashfs into the B partition, we can shrink the size
of the initial image by half and also cut the flashing time in half.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: fixing unrelated change to home label - credits go to Quirin]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>