refactor: use imagetypes for swu generation
This patch reworks the implementation of the swupdate type.
All generic aspects are moved from the swupdate.inc file into the
swupdate class and made conditional on the swu type.
The sw-description file is now referenced using the image-type
infrastructure, which avoids manual additions to FILESEXTRAPATHS and
accidental overwrites of SRC_URI. The templating logic is moved into the
generic one provided by imagetypes.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: defer WIC_DEPLOY_PARTITIONS removal]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>