add linux 6.1 mainline kernel
This patch adds the vanilla 6.1 linux kernel for testing and for boards
which need a more recent kernel than what CIP can offer.
The kernel can be selected from the menu, similar to the other kernels.
As defconfig, we always use the kernels internal configuration, because
there is no cip-kernel-config for that version yet. In addition, the
squashfs knob is selected to support the swupdate pattern without any
modifications.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: Update to latest 6.1.15, adjust copyright header]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>