add missing dependency to template subst in swu
Before creating swu files, the sw-description template needs to be
expanded. This logic is also independent of the imagetype.
Previously, only some image types had this task in the dependency chain
before invoking the swu creation. By that, it was working for e.g. wic,
but not for ext4 rootfs images (referring to the type of the abrootfs).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>