disable compression of squashfs swu artifact
The squashfs itself is already compressed in the default configuration.
By that, an additional compression of the artifact put into the .swu
file is not required. By disabling this compression, less temporary
space is required when flashing the image onto the target and also the
overhead of the decompression is avoided.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>