add support to limit resource consumption of squashfs creation
This patch makes it possible to control the number of CPUs and the
memory usage when generating a squashfs. The variable names used
follow the imagetype variable scheme.

Especially when compressing a squashfs internally with cpu and memory
heavy compressors like xz, we inherit the issues we had with the image
compression in the past. On CI systems, where many tasks might run in
parallel, this often led to OOMs and poor performance. By providing
these options similiar to the other imagetype options, we give CI
administrators an easy way to control the memory consumption,
independent of the other options passed to mksquashfs.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>