wic: Use INITRD_DEPLOY_FILE instead IMAGE_INITRD
With commit https://github.com/ilbers/isar/commit/f0608683fd18f96319332f2ea3827536bd2fd761 the variable
INITRD_DEPLOY_FILE should be used for creating the wic image,  see [1].

The current code breaks the build (wic) when using EBG in combination
with the stock initrd from the image in downstream layers. In this case,
the INITRD_IMAGE variable is empty, hence wic with EBG plugin cannot
find the initrd and fails. In cip-core, we always use a custom initrd
(by that INITRD_IMAGE is correctly defined), which hides the problem.

[1]: https://github.com/ilbers/isar/blob/master/RECIPE-API-CHANGELOG.md#working-with-a-custom-initramfs

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
[Jan: rebased and folded in impact description]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>