swupdate: Add recipe for 2022.12 for bullseye and bookworm
This allows modified swupdate builds in cip-core. It also
adds the option to use suricatta-lua instead of the default
suricatta-hawkbit to communicate with a webserver.

`suricatta-lua` is currently in use for internal projects. Some of these
projects are planned to be released as open source.

To enable suricatta-lua add the following line to the swupdate recipe or
use a bbappend.
```
DEB_BUILD_PROFILES += "pkg.swupdate.suricattalua"
```

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>