fix(start-efishell): argument passing to qemu
In commit 0129baec findings by shellcheck have been adressed. However,
this broke the QEMU arguments, as these do not allow to pass quoted
strings. By that, the deployment of secoure boot keys was broken.

To avoid similar issues in the future, we revert the code, disable the
shellcheck warning and add a note on the reason. In addition, the
format / driver of the disk is specified explicitly to avoid a warning
during startup.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>