squashfs: Fix vardeps for SQUASHFS_THREADS and SQUASHFS_MEMLIMIT
While SQUASHFS_DEFAULTS was decoupled from its potentially unstable
inputs, SQUASHFS_THREADS and SQUASHFS_MEMLIMIT, the consumer of
SQUASHFS_DEFAULTS was not. Rework this moving the vardepsexclude to
IMAGE_CMD:squashfs. This also obsoletes the related vardepvalue and
vardepsexclude for SQUASHFS_THREADS and SQUASHFS_MEMLIMIT.

As we are now expecting only generation limits via that ignored
variable, rename it to SQUASHFS_CREATION_LIMITS and deny overwriting.
Only SQUASHFS_THREADS and SQUASHFS_MEMLIMIT are supposed to be tuned
to the build environment - if at all.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>