kas/opt/security.yml : Change password to fit pam_passwdqc policy
Since we using pam_passwdqc from bookworm, our current security image login password
does not meet the requirements based on the below given pam_passwdqc policy.

When calculating the number of character classes, upper-case letters used as the
first character and digits used as the last character of a password are not counted.
This makes our password to have only 3 character classes. Hence we are adding multiple
capital characters to make the total number of character classes as 4.

Signed-off-by: Sai <Sai.Sathujoda@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>