make method to find /var mount configurable
Previously, the /var mount is only located by the partition label. This
is problematic, as collisions on this label might happen when working
with multiple disks. This patch makes the method to find the /var
partition configurable. By that, downstream implementations can use UUID
or PARTUUID based mounting.

For backwards compatibility we keep the current label-based mounting as
default.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>