verity: Prepare generation of metadata file for Isar update
This will be corrupted once SUDO_CHROOT is switched by isar to use
image_run which will also out some schroot boot-up messages.

Rather than deleting the metadata initially, we truncate any existing
file to 0 so that file permissions are set outside of the schroot
already. That is needed to modify the file after the schroot as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>