Update Isar revision
This brings a couple of fixes some downstream users are interesting it.
Therefore, make sure isar-cip-core itself is ready for it.

One adaption is needed because of the factored-out and modified
dpkg-source job. It now filters out binaries from the source package by
default, but we have some in edk2-standalonemm-rpmb. So, drop that
default filter in that package.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>