Migrate RISC-V support to Debian sid
With the promotion of riscv64 to an official port, building from
sid+sid-ports became broken long ago. Then cross-building from sid was
broken, but this has been fixed recently, and the release team is also
trying to keep it working. So it's time for us to migrate our
qemu-riscv64 to this new configuration.

We keep the "allow_failure" in CI as we are still building against
unstable, though. That may change once things move on into trixie.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>