u-boot: Update to 2024.01
Allows to drop the riscv patch.

This version switches us to the bootflow mechanism. The boot command is
simpler, but the secure config needs extra care to turn off all non-EFI
flows while keeping all switches that we need and that normally come via
too brought CONFIG_BOOTSTD_DEFAULTS.

u-boot-common is actually version-specific as it carries the sha.
Reflect that in the file name. Using ${PV} for its inclusion only works
directly, not via u-boot-qemu-common.inc, so do that from the respective
.bb files.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>