.gitlab-ci.yml: Run cve-checks job only when it is manually triggered in the pipeline
The cve-checks job generates CVE reports from latest copy of dpkg status files
using cve_checker.py script in debian-cve-checker. This job can only be triggered
manually with no dependency on build jobs running in build stage. This dependency
is removed so that one does not have to wait until all the jobs running in build
stage are passed or is unable to run the cve-checks at all because some job has
failed in the build stage.

Signed-off-by: Sai Sathujoda <Sai.Sathujoda@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>