.gitlab-ci.yml: Add condition check in .build_base to build v2 image
To re-use the built artifacts of v1 image, the consecutive build must
happen in the project directory of the job. So if build_swu_v2 is enabled,
then a consecutive build with a minor recipe change is done to get an .swu
artifact for swupdate testing in LAVA. deploy_file variable is default set
to wic so that after checking presence of .wic in the deploy script, all
the artifacts like wic, firmware etc. shall be deployed except .swu file.
Only .swu file will pushed when "swu" is passed to deploy script.

Signed-off-by: Sai Sathujoda <Sai.Sathujoda@toshiba-tsip.com>
[Jan: drop deploy_file variable]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>