swupdate images: initially deploy empty var part
Previously to this patch, the /var partition is only populated on the
initial image deployment. Further updates via swupdate do not update the
data on the partition. This is problematic, as issues w.r.t. missing
directories or packages that do not support this scheme are only
detected in the field or during update testing.

To avoid these late failures, we now always deploy an empty /var
partition. By that, potential issues can be spotted without the need to
run an update.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>