swupdate: check output of sign-swu
Check for signing errors to avoid an unusable swu file.

This also moves the siging out of the loop to generate
the cpio archive *.swu as the Messages from the signing
can lead to errors in the archive generation. The cpio
options are no longer using the short form.

Use local variables to increase readability.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
[Jan: small indention adjustments]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>