use ISAR snapshot infrastructure for sid
To have a stable base for debian sid, we need to pin the package set to
a fixed (known to be good) timestamp. Previously, this was done by
changing the apt sources list. As ISAR now supports to use snapshot
mirrors on-the-fly, we replace the fixed list with that logic.

This also makes it possible to later manually update the images (only
the non-immutable ones) from a running system, as the default apt
sources are not modified.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>