Refactor tar.gz image option
We have wic-targz.img, we have targz-img - that'c confusion. What we
need is a way to say "generate (also) tarball of the image" so that we
can feed LAVA specifically. If the normal target image is wic or already
a raw partition (simple QEMU targets) should not matter.

So, consolidate the kas options and the kconfig menu. Furthermore,
simplify the CI options so that tarballs are by default on and turn it
off actively in case that format has no user or cannot work in the first
place.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>