templates/swupdate_template.yml: Fail the job immediately after unexpected outcome
Previously to make the job incomplete, a sleep command is added in the test job under
the intention that it will cause a timeout error and make the job incomplete. Sometimes
due to unknown circumstances the error cannot be caught using this method. This also
increases the job time. To avoid this, "lava-test-raise <message>" can be used to
immediately fail the job as soon as unexpected outcome occurs.

Signed-off-by: sai ashrith sathujoda <sai.sathujoda@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>