ci: Require IEC tests to succeed as well
With the modifed failure detection scheme, those tests are no longer
reported as failure as they do (yet) let the job fail when a testcase
fails. Now that we pinned the test descriptions, we can also rely on
this property and enforce that the lava job and, thus, also the
respective gitlab-ci job succeed.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>