initramfs-overlay-hook: Add INITRAMFS_OVERLAY_MOUNT_OPTION
This allows to set options for the mount of the overlay storage device.
For security reasons, the mount options nodev, nosuid and noexec are
set.
As /var is not intended to support devices, nodev is set to ensure that
users cannot create a block or character special devices in /var.
Since /var is only intended for variable files such as logs, nosuid is
set to ensure that users cannot create setuid files in /var.
Noexec is set to ensure that users cannot run executable binaries.

Signed-off-by: Clara Kowalsky <clara.kowalsky@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>