submit_lava.sh: Modify environment slug in SQUAD
This commit modifies environment slug for test results hosted in SQUAD.
Currently the slug includes device name which keeps varying as that depends
on assigned device in LAVA Lab. Ex: "qemu-cip-siemens-muc", "qemu-renesas-01"
etc.

Environment slug should denote the test suite as well as the target on which
tests are executed. So it needs target information as part of it instead of
device allocated. This also helps in fetching the results from SQUAD using
squad-client while generating test reports.

Signed-off-by: sai ashrith sathujoda <sai.sathujoda@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>