fix(PREEMPT_RT): disable CONFIG_RT_GROUP_SCHED
This feature is simply broken on RT kernels that use cgroup2, as the
cpu controller features rt_period_us and rt_runtime_us are not
exposed, which are needed to configure things in a way that allows to
run tasks under RT/DL scheduling policies. It further seems to be
incompatible with the nohz_full option.

This patch just disables the feature, similar to upstream Debian.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>