swupdate: Perform checks for supported configurations
The upcoming trixie release is currently containing everything we need
from SWUpdate, so we are retiring the self-building option from that
release on - until someone comes up with a valid use case to restore it.
This happens to avoid exposing a configuration that is technically
selectable but became silently broken due to being effectively unused.

The Debian package enforces SWU signing. Therefore warn if a user
configures this out while using the upstream package.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>