initramfs-crypt-hook: Add support for expanding encrypted partition
The crypt hook already performs resizing or appropriately sized
formatting of partitions it handles during first boot. If we also add
the partition expansion logic from isar's expand-on-first-boot, we
get feature-complete expansion of encrypted partitions.

The feature is controlled by appending ":expand" to the affected
partition in CRYPT_PARTITIONS. In contrast to isar's expansion logic,
we do not have enough information at that time when it is called in
order to identify the target partition automatically. That is another
reason to embed this feature into the crypt hook.

There is not too much code to be taken from isar for this, so it is
easiest to duplicate and adjust that as needed, rather than introducing
some reusable upstream. This might be revisited in the future.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>