qemu-arm: Roll back to qemu_arm_defconfig
As long as the root cause of losing CONFIG_AUDIT while generating the
merged defconfig is not fully understood and resolved, move back to the
known-to-work qemu-specific config. That unbreaks our IEC test case [2].

There was a second issue, affecting the qemu_arm_defconfig directly. To
fix that, we also need to update the cip-kernel-config revision along
this.

[1] https://lore.kernel.org/cip-dev/5ed0fb0d-b505-4155-b330-069881cd9072@siemens.com/
[2] https://gitlab.com/cip-project/cip-core/isar-cip-core/-/jobs/8580918327

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>