erofs: ensure only dir prefixes are excluded
The regex used to implement the path exclusions in erofs needs to be
"anchored" to only match path prefixes but not arbitrary sections of the
path. Previously, also paths that contained an excluded part were
excluded as well. Further, we must single-quote the regex to ensure that
the '*' is not expanded by bash-globbing.

I further considered to use --exclude-path, but this is not an option as
it excludes the excluded directory as well. In case of /var and /boot
the directories (not the content) needs to be present as mountpoint.
These semantics differ from the equally named parameter in squashfs.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>