define RO_ROOTFS_TYPE before using
When including isar-cip-core in another layer and not using the cip-core
distribution, the RO_ROOTFS_TYPE is undefined. This results in parsing
errors.

We now provide a weak default in the verity and squashfs recipes to not
break the parsing in downstream layers. As we no longer need the
definition in the distro file, we remove it there.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>