kas/opt: add new option for security testing
Lava generates a directory, in LAVA called overlay, which contains
all scripts and tests of the test stage. The device-under-test
needs to be instrumented with this overlay.  LAVA provides the
possibility to download the overlay via http or NFS. We use curl
to download the overlay from a http server onto the target.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>