.gitlab-ci.yml: Deploy seperate home partition for qemu security targets
IEC layer tests run with /home as working directory. aide is unable to
detect changes made to a temporary file created under /home which is
now symlinked to /var/home. This resulted in failure of the following test cases
TC_CR3.4-RE2_1, TC_CR3.4_1, TC_CR6.2_1.

This is temporary change until it is understood if aide is oversensitive
here, and there is a way to make it accept our relocation.

Signed-off-by: sai ashrith sathujoda <sai.sathujoda@toshiba-tsip.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>