ebg-efi: add support to deploy additional files on boot partition
Some embedded systems (like the RPi 4b) cannot directly boot EFI.
Instead, they use a staged approach where a first stage loader then
loads the EFI provider, which then boots the efi target. For that, both
the first-stage loader, as well as additional config files need to be
added to the initial boot partition.

The bootimg-efi-isar already supports this for both grub and
systemd-boot by adding files to the IMAGE_EFI_BOOT_FILES bitbake
variable. We now add the same support to the EBG wic plugin, by copying
the code 1:1 to efibootguard-efi.py. We intentionally do not support
the IMAGE_EFI_BOOT_FILES_* as these are neither documented, not exposed
via the WIC_VARS.

Note, that by design these files are only copied to the ebg boot
partition, but not to the A/B kernel partitions. While we are not yet
EFI bootet, only a single partition is scanned by the firmware loader.
Once we bootet EFI, there is no need for the additional files anymore.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>