immutable-home: create relative symlink in rootfs_postprocess
The Debian policy 10.5 states that symlinks within the top-level
directory shall be realtive. Further, symlinks shall not be dead at any
point in time.

We now make the symlink relative and create the target directory. While
changing this, we also drop the unnecessary chrooting and make the path
to IMMUTABLE_DATA_DIR canonical.

Fixes: 786ddaa ("Move content of home to IMMUTABLE_DATA_DIR")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>