wic(ebg): add support to add verity env to cmdline
Passing the verity data via the initrd has proven problematic, as this
requires to rebuild the initrd on each (bitwise) change of the verity
container. For use cases with not bit-by-bit reproducible rootfs this
can lead to verity hash inconsistencies as the bitbake state diverges
from the file state in case some task artifacts are taken from the
SState cache. Further, the build time is prolonged as on every rootfs
change also the initfs needs to be rebuild (also polluting the SState
cache with not-reusable entries).

We now change this by adding the verity data to the kernel cmdline
(similar to how systemd envisions this). The ebg-boot wic plugin
already adds information about the imaged partitions to the kernel
cmdline of the UKI (or config). We now add support to add the verity
environment by setting the source parameter "verity_root=y". The
environment is read from the images *.verity.env file, converted
into the systemd-veritysetup-generator syntax and added to the
command line. As we currently do not use the systemd integration, we
use the parameter "cip.verity_root_options" instead. Later on this
can be replaced by "systemd.verity_root_options".

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Quirin: add IMAGE_BASENAME to search for correct verity.env ]
Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>