port verity env handling to new kernel cmdline infrastructure
Previously the verity env was injected as file into the initrd and then
sourced at runtime. As we now have infrastructure in place to add the
verity env to the kernel cmdline, we replace the hard-coded values with
logic to extract it from the kernel cmdline. By that, we also remove the
dependency between the initramfs build and the verity creation, which
reduces the set of dependencies (good for caching), as well as increases
the parallelism of the build.

Once widely supported, the initrd scripts can be completely replaced by
the systemd-veritysetup-generator (identical syntax).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Quirin: fix typos in commit message]
Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>