port abrootfs env handling to new kernel cmdline infrastructure
Previously the abrootfs env was injected as file into the initrd and then
sourced at runtime. As we now have infrastructure in place to add the
abrootfs env to the kernel cmdline, we replace the hard-coded values with
logic to extract it from the kernel cmdline. By that, we also remove the
dependency between the initramfs build and the abrootfs creation, which
reduces the set of dependencies (good for caching), as well as increases
the parallelism of the build.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
[Jan: dropped debug left-over "set -x", dropped ABROOTFS_IMAGE_RECIPE from security.yml]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>