linux-cip: Fold rt-specific include into a single file
This will simplify the addition of the 6.12 kernel where there is no
need for a separate recipe as there is no separate code base anymore.

At this chance, it turned out that we never built the 6.1-cip-rt kernel
in isar-cip-core with RT settings. That is automatically fixed now as
well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>