Add support for CIP kernel 6.12
One difference with 6.12 is that there are no longer two different code
bases for vanilla and RT. Therefore, simply establish a link from the RT
recipe to the regular one so that we can continue to control the
activation of RT kernel config by setting the KERNEL_NAME to "cip-rt".

As trixie upstream is heading for 6.12, lift our default for CIP images
built for it accordingly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>