.gitlab-ci.yml: Increase IEC test timeout for real hardware(M-COM).
Sometimes the current timeout of 60 mins for IEC tests execution for M-COM
device on LAVA lab is not sufficient. The gitlab CI job fails due to timeout
issue. Increase timeout to 120 min to avoid CI job failures.

It's a temporary fix, so need to explore the reason for the longer runtime
in LAVA setup for M-COM.

[1] https://lava.ciplatform.org/scheduler/job/1281301

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>