initramfs-crypt-hook: store initial encryption key in TPM2
cryptsetup and systemd-cryptenroll do not currently support encrypting
partitions directly with keys stored in the TPM. For that reason a
temporary random password is used for the initial encryption, which will
get removed once the TPM2 token is added to the partition.

However this process does not allow continuing with the encryption
process in case of power failure.

For that reason store the initial encryption password in the TPM,
protected via measured boot based on the TPM2-PCR:7 register, which is
the default for `systemd-cryptenroll` and load it if it exists.

We only provide this for bullseye and newer because tpm2-tools version
3.x as provided with buster makes it impractical to support this
feature.

Signed-off-by: Claudius Heine <ch@denx.de>
[Jan: rebased, added buster stubs]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>